refactor(agent-toolkit): move publish_workflow gql to graphql.dev file, fix draft ID note in description
