+++ packages/agent-toolkit/codegen.yml
@@ -18,7 +18,10 @@ projects:
                 - node -e "const fs = require('fs'); fs.writeFileSync('src/monday-graphql/generated/graphql/index.ts', '/* eslint-disable */\nexport * from \'./gql\';');"
   dev:
     schema: 'src/monday-graphql/schema.dev.graphql'
-    documents: 'src/**/*.graphql.dev.ts'
+    documents:
+      - 'src/**/*.graphql.dev.ts'
+      # TODO: remove exclusion once publish_workflow lands in the dev schema (automation-builders#3085)
+      - '!src/**/publish-workflow.graphql.dev.ts'
     ignoreNoDocuments: true
     extensions:
       codegen:
