+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflow-builder-tools/publish-workflow/publish-workflow.graphql.dev.ts
@@ -1,0 +1,14 @@
+import { gql } from 'graphql-request';
+
+export const publishWorkflowMutation = gql`
+  mutation publishWorkflow($workflow_object_id: ID!, $workflow_draft_id: ID!, $should_activate: Boolean) {
+    publish_workflow(
+      workflow_object_id: $workflow_object_id
+      workflow_draft_id: $workflow_draft_id
+      should_activate: $should_activate
+    ) {
+      workflow_object_id
+      workflow_live_id
+    }
+  }
+`;
